pc - move UCSBAPIQuarter from models to entities and add repository
